tests/category: Move tests into `routes` submodules
