+++ src/tests/routes/category_slugs/mod.rs
@@ -1,0 +1,1 @@
+pub mod list;
